Classifier test still failing but...
have used composition rather than inheriting from sklearn.Pipeline
for the BuildClassifier class.
This is so I can add a evaluate and peform method to the class.
This is so I can overcome needing to a) override any method and
b) encountering problems with using the pipeline attribute of BuildClassifier.
It does mean that BuildClassifier will need to be initiated with the steps parameter.