Discovered error with test_tfidf_weighting
Test was failing because was assertEqual is not able to test arrays
Further investigation revealed that the tfidf_fit_transform is not
carrying out the transform but only the the fit and then trying to
return it.
Perhaps it should not return but just carry out the transform
Then self.tf_transformer would be accessed directly after.
How can I stop the self.tf_transformer variable being accessed
before the fit (i.e. learning) has carried out. Should I worry?