Created clf.pickle.
clf.pickle is a MultinomialNB class fitted to test data that
has been fitted and transformed with CountVectorizer and
TfidfTransformer.

The web app will load this pickle as a model into which
user data can be predicted against, after user data has been
transformed with CV and TF.

Evaluating the performance of clf.pickle has not been
carried out.