Able to classify user input but not display it.
THis is becauase of a TypeError and the array not being JSON seriablizable
Used flasks interactive python debugger extensively this evening rather
than writing tests.

To get to this stage the classifier, count_vect and tf_transformer all
had to be re-pickled from predictocite\main.py as was getting dimension
mismatches and feature mismatch errors.